requirements.txt
# Report Processing System - Requirements

# Web Framework
flask>=2.3.0
flask-cors>=4.0.0
flask-sqlalchemy>=3.0.0
flask-migrate>=4.0.0
flask-jwt-extended>=4.5.0
gunicorn>=21.0.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0  # PostgreSQL adapter
alembic>=1.10.0

# Document Processing
python-docx>=0.8.11
PyPDF2>=3.0.1
pdfplumber>=0.10.2

# Audio Processing
SpeechRecognition>=3.10.0
pydub>=0.25.1

# API & External Services
openai>=1.3.0
requests>=2.31.0

# File Handling
python-magic>=0.4.27
watchdog>=3.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Authentication & Security
werkzeug>=2.3.0
bcrypt>=4.0.0

# Async Processing
celery>=5.3.0
redis>=4.5.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
rich>=13.0.0